+++ include/envoy/server/options.h
@@ -123,9 +123,9 @@ public:
   virtual const std::string& logFormat() const PURE;
 
   /**
-   * @return const bool indicating whether to escape newline characters in logs.
+   * @return const bool indicating whether to escape c-style escape sequences in logs.
    */
-  virtual bool logFormatEscapeNewlines() const PURE;
+  virtual bool logFormatEscaped() const PURE;
 
   /**
    * @return const std::string& the log file path.