+++ source/server/options_impl.cc
@@ -72,8 +72,8 @@ OptionsImpl::OptionsImpl(int argc, const char* const* argv,
       "", "component-log-level", component_log_level_string, false, "", "string", cmd);
   TCLAP::ValueArg<std::string> log_format("", "log-format", log_format_string, false,
                                           Logger::Logger::DEFAULT_LOG_FORMAT, "string", cmd);
-  TCLAP::SwitchArg log_format_escape_newlines("", "log-format-escape-newlines",
-                                              "Escape newline characters in the application logs",
+  TCLAP::SwitchArg log_format_escaped("", "log-format-escaped",
+                                              "Escape c-style escape sequences in the application logs",
                                               cmd, false);
   TCLAP::ValueArg<std::string> log_path("", "log-path", "Path to logfile", false, "", "string",
                                         cmd);
@@ -152,7 +152,7 @@ OptionsImpl::OptionsImpl(int argc, const char* const* argv,
   }
 
   log_format_ = log_format.getValue();
-  log_format_escape_newlines_ = log_format_escape_newlines.getValue();
+  log_format_escaped_ = log_format_escaped.getValue();
 
   parseComponentLogLevels(component_log_level.getValue());
 
@@ -268,7 +268,7 @@ Server::CommandLineOptionsPtr OptionsImpl::toCommandLineOptions() const {
   command_line_options->set_log_level(spdlog::level::to_string_view(logLevel()).data(),
                                       spdlog::level::to_string_view(logLevel()).size());
   command_line_options->set_log_format(logFormat());
-  command_line_options->set_log_format_escape_newlines(logFormatEscapeNewlines());
+  command_line_options->set_log_format_escaped(logFormatEscaped());
   command_line_options->set_log_path(logPath());
   command_line_options->set_service_cluster(serviceClusterName());
   command_line_options->set_service_node(serviceNodeName());
@@ -304,7 +304,7 @@ OptionsImpl::OptionsImpl(const std::string& service_cluster, const std::string& 
                          const std::string& service_zone, spdlog::level::level_enum log_level)
     : base_id_(0u), concurrency_(1u), config_path_(""), config_yaml_(""),
       local_address_ip_version_(Network::Address::IpVersion::v4), log_level_(log_level),
-      log_format_(Logger::Logger::DEFAULT_LOG_FORMAT), log_format_escape_newlines_(false),
+      log_format_(Logger::Logger::DEFAULT_LOG_FORMAT), log_format_escaped_(false),
       restart_epoch_(0u), service_cluster_(service_cluster), service_node_(service_node),
       service_zone_(service_zone), file_flush_interval_msec_(10000), drain_time_(600),
       parent_shutdown_time_(900), mode_(Server::Mode::Serve), hot_restart_disabled_(false),